apps/api/requirements.txt
# Web framework + server
fastapi==0.135.3
uvicorn[standard]==0.44.0

# Data validation (FastAPI already pulls it, but pin explicitly since we import directly)
pydantic==2.13.0

# Database + auth
supabase==2.28.3

# LLM providers
anthropic==0.95.0
openai==2.32.0

# Vector store
qdrant-client==1.17.1

# Environment variable loading
python-dotenv==1.2.2

# HTTP client (used directly in some RAG code paths)
httpx==0.28.1

# Testing
pytest==9.0.3
pytest-asyncio==1.3.0

# Evals
deepeval==3.9.7
